+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.4.0
 
 - If an `ActionController::Parameters` is passed as parameters these will automatically marked as permitted
+- Added `attributes_from_user` method to return only the changed attributes
 
 ## v1.3.0